requirements.txt
customtkinter
tkinterdnd2
pycryptodome
pillow